Create synapse module with api and workload modules inside
