+++ src/synapse/api.py
@@ -46,70 +46,6 @@ class NetworkError(SynapseAPIError):
     """Exception raised when requesting API fails due network issues."""
 
 
-def _get_nonce() -> str:
-    """Get nonce.
-
-    Returns:
-        The nonce returned by Synapse API.
-
-    Raises:
-        NetworkError: if there was an error fetching the nonce.
-    """
-    try:
-        res = requests.get(REGISTER_URL, timeout=5)
-        res.raise_for_status()
-        return res.json()["nonce"]
-    except (
-        requests.exceptions.ConnectionError,
-        requests.exceptions.Timeout,
-        requests.exceptions.HTTPError,
-    ) as exc:
-        logger.error("Failed to request %s : %s", REGISTER_URL, exc)
-        raise NetworkError(f"Failed to request {REGISTER_URL}.") from exc
-
-
-def _generate_mac(
-    shared_secret: str,
-    nonce: str,
-    user: str,
-    password: str,
-    admin: bool = False,
-    user_type: typing.Optional[str] = None,
-) -> str:
-    """Generate mac to register user.
-
-    "The MAC is the hex digest output of the HMAC-SHA1 algorithm, with the key being the shared
-    secret and the content being the nonce, user, password, either the string "admin" or "notadmin"
-    , and optionally the user_type each separated by NULs.".
-    Extracted from: https://matrix-org.github.io/synapse/latest/admin_api/register_api.html
-
-    Args:
-        shared_secret: registration_shared_secret from configuration file.
-        nonce: nonce generated.
-        user: username.
-        password: password.
-        admin: if is admin. Default False.
-        user_type: user type. Defaults to None.
-
-    Returns:
-        User in HMAC format as a string of hexadecimals.
-        This format is expected by the Synapse API.
-    """
-    mac = hmac.new(key=shared_secret.encode("utf8"), digestmod=hashlib.sha1)
-    mac.update(nonce.encode("utf8"))
-    mac.update(b"\x00")
-    mac.update(user.encode("utf8"))
-    mac.update(b"\x00")
-    mac.update(password.encode("utf8"))
-    mac.update(b"\x00")
-    mac.update(b"admin" if admin else b"notadmin")
-    if user_type:
-        mac.update(b"\x00")
-        mac.update(user_type.encode("utf8"))
-
-    return mac.hexdigest()
-
-
 def register_user(registration_shared_secret: str, user: User) -> None:
     """Register user.
 
@@ -142,6 +78,70 @@ def register_user(registration_shared_secret: str, user: User) -> None:
     try:
         res = requests.post(REGISTER_URL, json=data, timeout=5)
         res.raise_for_status()
+    except (
+        requests.exceptions.ConnectionError,
+        requests.exceptions.Timeout,
+        requests.exceptions.HTTPError,
+    ) as exc:
+        logger.error("Failed to request %s : %s", REGISTER_URL, exc)
+        raise NetworkError(f"Failed to request {REGISTER_URL}.") from exc
+
+
+def _generate_mac(
+    shared_secret: str,
+    nonce: str,
+    user: str,
+    password: str,
+    admin: bool = False,
+    user_type: typing.Optional[str] = None,
+) -> str:
+    """Generate mac to register user.
+
+    "The MAC is the hex digest output of the HMAC-SHA1 algorithm, with the key being the shared
+    secret and the content being the nonce, user, password, either the string "admin" or "notadmin"
+    , and optionally the user_type each separated by NULs.".
+    Extracted from: https://matrix-org.github.io/synapse/latest/admin_api/register_api.html
+
+    Args:
+        shared_secret: registration_shared_secret from configuration file.
+        nonce: nonce generated.
+        user: username for the new user.
+        password: password used for authentication.
+        admin: if is admin. Default False.
+        user_type: user type. Defaults to None.
+
+    Returns:
+        User in HMAC format as a string of hexadecimals.
+        This format is expected by the Synapse API.
+    """
+    mac = hmac.new(key=shared_secret.encode("utf8"), digestmod=hashlib.sha1)
+    mac.update(nonce.encode("utf8"))
+    mac.update(b"\x00")
+    mac.update(user.encode("utf8"))
+    mac.update(b"\x00")
+    mac.update(password.encode("utf8"))
+    mac.update(b"\x00")
+    mac.update(b"admin" if admin else b"notadmin")
+    if user_type:
+        mac.update(b"\x00")
+        mac.update(user_type.encode("utf8"))
+
+    return mac.hexdigest()
+
+
+def _get_nonce() -> str:
+    """Get nonce.
+
+    Returns:
+        The nonce returned by Synapse API.
+
+    Raises:
+        NetworkError: if there was an error fetching the nonce.
+    """
+    try:
+        res = requests.get(REGISTER_URL, timeout=5)
+        res.raise_for_status()
+        return res.json()["nonce"]
     except (
         requests.exceptions.ConnectionError,
         requests.exceptions.Timeout,
