+++ tests/unit/test_database.py
@@ -14,6 +14,7 @@ from ops.testing import Harness
 from psycopg2 import sql
 
 import database_observer
+import synapse
 from charm_types import DatasourcePostgreSQL
 from constants import SYNAPSE_CONTAINER_NAME
 from database_client import DatabaseClient
@@ -190,7 +191,7 @@ def test_relation_as_datasource(
     )
     assert expected == harness.charm.database.get_relation_as_datasource()
     assert harness.charm.app.name == harness.charm.database.get_database_name()
-    synapse_env = harness.charm._synapse.synapse_environment()
+    synapse_env = synapse.get_environment(harness.charm._charm_state)
     assert synapse_env["POSTGRES_DB"] == expected["db"]
     assert synapse_env["POSTGRES_HOST"] == expected["host"]
     assert synapse_env["POSTGRES_PORT"] == expected["port"]
